split purchase_ext_sst
