+++ purchase_delivery_staff/README.rst
@@ -1,0 +1,49 @@
+=======================
+Purchase Delivery Staff
+=======================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/github-qrtl%2Fsst--custom-lightgray.png?logo=github
+    :target: https://github.com/qrtl/sst-custom/tree/11.0/purchase_delivery_staff
+    :alt: qrtl/sst-custom
+
+|badge1| |badge2| 
+
+This module add shop_id and purchase_by_id in purchase order.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/qrtl/sst-custom/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/qrtl/sst-custom/issues/new?body=module:%20purchase_delivery_staff%0Aversion:%2011.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Quartile Limited
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `qrtl/sst-custom <https://github.com/qrtl/sst-custom/tree/11.0/purchase_delivery_staff>`_ project on GitHub.
+
+You are welcome to contribute.
