+++ purchase_delivery_staff/models/__init__.py
@@ -1,0 +1,2 @@
+from . import hr_employee
+from . import purchase_order
