+++ purchase_delivery_staff/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module add shop_id and purchase_by_id in purchase order.
