+++ purchase_ext_sst/__manifest__.py
@@ -7,16 +7,10 @@
     "website": "https://www.quartile.co",
     "category": "Purchase",
     "license": "LGPL-3",
-    "depends": ["hr", "purchase", "purchase_convert_sale"],
+    "depends": ["hr", "purchase_convert_sale"],
     "data": [
         "security/ir.model.access.csv",
-        "data/request_channel_data.xml",
-        "data/purchase_category_data.xml",
-        "views/request_channel_views.xml",
-        "views/request_medium_views.xml",
-        "views/purchase_category_views.xml",
         "views/purchase_order_views.xml",
-        "views/hr_employee_views.xml",
     ],
     "installable": True,
 }
