+++ purchase_ext_sst/models/__init__.py
@@ -1,7 +1,3 @@
-from . import request_channel
-from . import request_medium
-from . import purchase_category
 from . import purchase_order_line_hook_onchange_product_id
 from . import purchase_order
 from . import purchase_order_line
-from . import hr_employee
