+++ purchase_ext_sst/models/purchase_order.py
@@ -8,34 +8,10 @@ from odoo.exceptions import UserError
 class PurchaseOrder(models.Model):
     _inherit = "purchase.order"
 
-    request_channel_id = fields.Many2one("request.channel", "Request Channel")
-    request_medium_id = fields.Many2one("request.medium", "Request Medium")
     employee_id = fields.Many2one("hr.employee", "Received By")
-    purchase_category_id = fields.Many2one("purchase.category", "Purchase Category",)
     call_back = fields.Boolean("Call Back")
-    shop_id = fields.Many2one("stock.warehouse", "Shop")
-    phone_update = fields.Char()
-    mobile_update = fields.Char()
-    phone_search = fields.Char(
-        states={
-            "purchase": [("readonly", True)],
-            "done": [("readonly", True)],
-            "cancel": [("readonly", True)],
-        },
-    )
-    supplier_phone = fields.Char(
-        related="partner_id.phone", string="Supplier Phone", readonly=True, store=True,
-    )
-    supplier_mobile = fields.Char(
-        related="partner_id.mobile",
-        string="Supplier Mobile",
-        readonly=True,
-        store=True,
-    )
-    tentative_name = fields.Char("Tentative Name")
     address = fields.Char()
     remark = fields.Text("Remark")
-    purchased_by_id = fields.Many2one("hr.employee", "Delivery Staff")
     worked_hours = fields.Selection(
         [
             (num, num + " hours")
@@ -64,53 +40,10 @@ class PurchaseOrder(models.Model):
         ],
         string="Worked Hours",
     )
-    date_planned = fields.Datetime(compute=False,)
+    date_planned = fields.Datetime(
+        compute=False,
+    )
     sale_prediction_amount = fields.Monetary("Sales Prediction")
-
-    @api.onchange("purchased_by_id")
-    def onchange_purchased_by_id(self):
-        if (not self.shop_id and self.purchased_by_id) or (
-            self.shop_id
-            and self.purchased_by_id
-            and self.purchased_by_id.shop_id != self.shop_id
-        ):
-            self.shop_id = self.purchased_by_id.shop_id
-
-    @api.onchange("shop_id")
-    def onchange_shop_id(self):
-        ids = []
-        if self.shop_id:
-            # Update domain filter on delivery staff
-            staffs = self.env["hr.employee"].search([("shop_id", "=", self.shop_id.id)])
-            ids.append(("id", "in", staffs.ids))
-            # Clear the delivery staff value
-            if (
-                self.purchased_by_id
-                and self.purchased_by_id.shop_id
-                and self.purchased_by_id.shop_id != self.shop_id
-            ):
-                self.purchased_by_id = False
-            # Update picking_type_id
-            picking_type_id = self.env["stock.picking.type"].search(
-                [
-                    ("code", "=", "incoming"),
-                    ("warehouse_id", "=", self.shop_id.id),
-                    (
-                        "warehouse_id.company_id",
-                        "in",
-                        [
-                            self.env.context.get(
-                                "company_id", self.env.user.company_id.id
-                            ),
-                            False,
-                        ],
-                    ),
-                ],
-                limit=1,
-            )
-            if picking_type_id:
-                self.picking_type_id = picking_type_id
-        return {"domain": {"purchased_by_id": ids}}
 
     @api.multi
     def open_record(self):
@@ -127,90 +60,6 @@ class PurchaseOrder(models.Model):
         }
 
     @api.multi
-    def write(self, vals):
-        # Propose supplier logic
-        if (
-            "phone_search" in vals
-            or "phone_update" in vals
-            or "mobile_update" in vals
-            or "tentative_name" in vals
-        ):
-            val_phone_search = "phone_search" in vals and vals["phone_search"] or False
-            val_phone = "phone_update" in vals and vals["phone_update"] or False
-            val_mobile = "mobile_update" in vals and vals["mobile_update"] or False
-            for order in self:
-                val_tent_name = (
-                    "tentative_name" in vals
-                    and vals["tentative_name"]
-                    or order.tentative_name
-                )
-                partner_id = (
-                    "partner_id" in vals and vals["partner_id"] or order.partner_id.id
-                )
-                if self.is_default_partner(partner_id) and (
-                    val_phone or val_mobile or val_phone_search
-                ):
-                    partner_id = self.create_partner(
-                        val_phone or val_phone_search, val_mobile
-                    ).id
-                if not self.is_default_partner(partner_id):
-                    partner = self.env["res.partner"].browse(partner_id)
-                    self.update_partner(partner, val_phone, val_mobile, val_tent_name)
-                vals["partner_id"] = partner_id
-            vals["phone_update"] = vals["mobile_update"] = False
-        res = super(PurchaseOrder, self).write(vals)
-        # Update product shop_id and purchase_by_id
-        if "order_line" in vals or "shop_id" in vals or "purchased_by_id" in vals:
-            for order in self:
-                for line in order.order_line:
-                    product = line.product_id
-                    if product.shop_id != order.shop_id:
-                        product.shop_id = order.shop_id
-                    if product.purchased_by_id != order.purchased_by_id:
-                        product.purchased_by_id = order.purchased_by_id
-        return res
-
-    @api.model
-    def create(self, vals):
-        val_phone_search = "phone_search" in vals and vals["phone_search"] or False
-        val_phone = "phone_update" in vals and vals["phone_update"] or False
-        val_mobile = "mobile_update" in vals and vals["mobile_update"] or False
-        val_tent_name = "tentative_name" in vals and vals["tentative_name"] or False
-        if self.is_default_partner(vals["partner_id"]) and (
-            val_phone or val_mobile or val_phone_search
-        ):
-            vals["partner_id"] = self.create_partner(
-                val_phone or val_phone_search, val_mobile
-            ).id
-        if not self.is_default_partner(vals["partner_id"]):
-            partner = self.env["res.partner"].browse(vals["partner_id"])
-            self.update_partner(partner, val_phone, val_mobile, val_tent_name)
-        vals["phone_update"] = vals["mobile_update"] = False
-        return super(PurchaseOrder, self).create(vals)
-
-    @api.onchange("phone_search")
-    def onchange_phone_search(self):
-        if self.phone_search:
-            self.phone_update = False
-            self.mobile_update = False
-            partner = self.get_partner_from_phone(self.phone_search)
-            if partner:
-                self.partner_id = partner
-            elif self.partner_id:
-                if self.is_default_partner(self.partner_id.id):
-                    self.phone_update = self.phone_search
-
-    @api.onchange("phone_update")
-    def onchange_phone_update(self):
-        if self.phone_update:
-            return self.check_onchange_phone(self.phone_update, "phone_update")
-
-    @api.onchange("mobile_update")
-    def onchange_mobile_update(self):
-        if self.mobile_update:
-            return self.check_onchange_phone(self.mobile_update, "mobile_update")
-
-    @api.multi
     def button_confirm(self):
         for purchase_order in self:
             if self.is_default_partner(purchase_order.partner_id.id):
@@ -222,76 +71,6 @@ class PurchaseOrder(models.Model):
                 for order_line in purchase_order.order_line:
                     order_line.date_planned = purchase_order.date_planned
         return super(PurchaseOrder, self).button_confirm()
-
-    def check_onchange_phone(self, phone, field):
-        try:
-            partner = self.get_partner_from_phone(phone)
-            if self.partner_id and partner and partner != self.partner_id:
-                conflicts_user = _("\n%s\n- Phone: %s\n- Mobile: %s\n") % (
-                    partner.name,
-                    partner.phone or "",
-                    partner.mobile or "",
-                )
-                return {
-                    "warning": {
-                        "message": _(
-                            "The entered phone (%s) conflicts with "
-                            "the following user(s):\n%s"
-                        )
-                        % (phone or "N/A", conflicts_user)
-                    },
-                    "value": {field: False},
-                }
-        except UserError as e:
-            return {"warning": {"message": e.name}, "value": {field: False}}
-
-    def get_partner_from_phone(self, phone):
-        Partner = self.env["res.partner"]
-        partners = False
-        # here we use "or" condition instead of "and"
-        if phone:
-            partners = Partner.search(
-                [
-                    "|",
-                    ("phone", "=", phone),
-                    ("mobile", "=", phone),
-                    ("active", "=", True),
-                    ("supplier", "=", True),
-                ]
-            )
-        if partners and len(partners) > 1:
-            conflicts_users_list = ""
-            for partner in partners:
-                conflicts_users_list += _("\n%s\n- Phone: %s\n- " "Mobile: %s\n") % (
-                    partner.name,
-                    partner.phone or "",
-                    partner.mobile or "",
-                )
-            raise UserError(
-                _("The entered phone (%s) " "conflicts with the following user(s):\n%s")
-                % (phone or "N/A", conflicts_users_list)
-            )
-        return partners if partners else False
-
-    def create_partner(self, phone, mobile):
-        partner = self.env["res.partner"].create(
-            {
-                "name": "未確認",
-                "phone": phone,
-                "mobile": mobile,
-                "supplier": True,
-                "customer": False,
-            }
-        )
-        return partner
-
-    def update_partner(self, partner, phone, mobile, tent_name):
-        if phone and partner.phone != phone:
-            partner.phone = phone
-        if mobile and partner.mobile != mobile:
-            partner.mobile = mobile
-        if tent_name and partner.name != tent_name:
-            partner.name = tent_name
 
     def is_default_partner(self, partner_id):
         company_id = self.env.user.company_id.id
