+++ purchase_ext_sst/models/purchase_order_line_hook_onchange_product_id.py
@@ -35,7 +35,8 @@ def onchange_product_id(self):
     }
 
     product_lang = self.product_id.with_context(
-        lang=self.partner_id.lang, partner_id=self.partner_id.id,
+        lang=self.partner_id.lang,
+        partner_id=self.partner_id.id,
     )
     self.name = product_lang.display_name
     if product_lang.description_purchase:
