--- purchase_ext_sst/views/hr_employee_views.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <record id="view_employee_form" model="ir.ui.view">
-        <field name="name">hr.employee.form</field>
-        <field name="model">hr.employee</field>
-        <field name="inherit_id" ref="hr.view_employee_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='department_id']" position="before">
-                <field name="shop_id" />
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
