+++ purchase_ext_sst/views/purchase_order_views.xml
@@ -23,7 +23,7 @@
             <xpath expr="//field[@name='product_id']" position="attributes">
                 <attribute
           name="context"
-        >{"default_evaluated_by_id": parent.employee_id, "default_purchase_category_id": parent.purchase_category_id}</attribute>
+        >{"default_evaluated_by_id": parent.employee_id}</attribute>
             </xpath>
             <xpath expr="//field[@name='product_id']" position="after">
                 <button
@@ -32,21 +32,6 @@
           icon="fa-sign-in"
           class="oe_read_only"
         />
-            </xpath>
-            <xpath expr="//field[@name='partner_id']" position="before">
-                <field name="phone_search" />
-            </xpath>
-            <xpath expr="//field[@name='partner_id']" position="after">
-                <field name="supplier_phone" />
-                <field name="supplier_mobile" />
-            </xpath>
-            <xpath expr="//field[@name='date_order']" position="after">
-                <field
-          name="tentative_name"
-          string="Supplier Name Update (Tentative Name)"
-        />
-                <field name="phone_update" />
-                <field name="mobile_update" />
             </xpath>
             <xpath expr="//sheet/group" position="after">
                 <group>
@@ -61,24 +46,17 @@
                 </group>
                 <group>
                     <group>
-                        <field name="request_channel_id" />
-                        <field name="request_medium_id" />
                         <field name="call_back" />
                         <field name="worked_hours" />
                     </group>
                     <group>
                         <field name="employee_id" />
-                        <field name="shop_id" />
-                        <field name="purchase_category_id" />
                     </group>
                 </group>
                 <group>
                     <field name="remark" />
                     <field name="sale_prediction_amount" />
                 </group>
-            </xpath>
-            <xpath expr="//field[@name='incoterm_id']" position="after">
-                <field name="purchased_by_id" />
             </xpath>
             <xpath
         expr="//notebook/page[1]//tree/field[@name='date_planned']"
@@ -95,7 +73,7 @@
         </field>
     </record>
 
-    <record id="purchase_convert_sale_user_id_purchase_order_form_inherit" model="ir.ui.view">
+    <record id="purchase_convert_sale_purchase_order_form_inherit" model="ir.ui.view">
         <field name="name">purchase.order.form.inherit</field>
         <field name="model">purchase.order</field>
         <field
@@ -106,14 +84,6 @@
             <xpath expr="//field[@name='user_id']" position="attributes">
                 <attribute name="invisible">True</attribute>
             </xpath>
-        </field>
-    </record>
-
-    <record id="purchase_convert_sale_sale_order_id_purchase_order_form_inherit" model="ir.ui.view">
-        <field name="name">purchase.order.form.inherit</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase_convert_sale.purchase_order_form_inherit" />
-        <field name="arch" type="xml">
             <xpath expr="//field[@name='sale_order_id']" position="attributes">
                 <attribute name="invisible">True</attribute>
             </xpath>
@@ -135,15 +105,10 @@
                 <attribute name="domain">[('supplier', '=', True)]</attribute>
             </xpath>
             <xpath expr="//field[@name='name']" position="after">
-                <field name="purchase_category_id" />
                 <field name="picking_type_id" invisible="1" />
-                <field name="shop_id" />
                 <field name="call_back" />
-                <field name="request_channel_id" />
-                <field name="request_medium_id" />
             </xpath>
             <xpath expr="//field[@name='partner_id']" position="after">
-                <field name="phone_search" />
                 <field name="address" />
                 <field name="remark" />
                 <field name="sale_prediction_amount" />
@@ -169,19 +134,11 @@
         <field name="inherit_id" ref="purchase.view_purchase_order_filter" />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='name']" position="before">
-                <field
-          name="supplier_phone"
-          filter_domain="['|', ('supplier_phone', 'ilike', self), ('supplier_mobile', 'ilike', self)]"
-        />
                 <field name="address" />
             </xpath>
             <xpath expr="//field[@name='partner_id']" position="after">
-                <field name="request_channel_id" />
-                <field name="request_medium_id" />
                 <field name="call_back" />
                 <field name="employee_id" />
-                <field name="shop_id" />
-                <field name="purchase_category_id" />
                 <field name="remark" />
             </xpath>
             <xpath expr="//filter[@name='to_approve']" position="after">
@@ -204,5 +161,11 @@
             </xpath>
         </field>
     </record>
-
+    <menuitem
+        id="menu_purchase_in_config_purchase"
+        name="Purchases"
+        parent="purchase.menu_purchase_config"
+        sequence="50"
+        groups="base.group_no_one"
+    />
 </odoo>
