+++ purchase_order_category/models/__init__.py
@@ -1,0 +1,2 @@
+from . import purchase_category
+from . import purchase_order
