+++ purchase_order_category/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module adds purchase_category_id in purchase order.
