+++ purchase_order_category/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_purchase_category_user,purchase_category_user,model_purchase_category,base.group_user,1,0,0,0
+access_purchase_category_manager,purchase_category_manager,model_purchase_category,purchase.group_purchase_manager,1,1,1,1
