+++ purchase_order_request_channel/data/request_channel_data.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo noupdate="1">
-
     <record id="request_channel_phone" model="request.channel">
         <field name="name">TEL</field>
         <field eval="False" name="company_id" />
@@ -17,5 +16,4 @@
         <field name="name">その他</field>
         <field eval="False" name="company_id" />
     </record>
-
 </odoo>
