+++ purchase_order_request_channel/models/__init__.py
@@ -1,0 +1,2 @@
+from . import request_channel
+from . import purchase_order
