+++ purchase_order_request_channel/models/request_channel.py
@@ -1,5 +1,5 @@
 # Copyright 2017 Quartile Limited
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import fields, models
