+++ purchase_order_request_channel/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module adds request_channel_id in purchase order.
