+++ purchase_order_request_channel/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_request_channel_user,request_channel_user,model_request_channel,base.group_user,1,0,0,0
+access_request_channel_manager,request_channel_manager,model_request_channel,purchase.group_purchase_manager,1,1,1,1
