+++ purchase_order_request_medium/models/purchase_order.py
@@ -1,0 +1,10 @@
+# Copyright 2017-2018 Quartile Limited
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class PurchaseOrder(models.Model):
+    _inherit = "purchase.order"
+
+    request_medium_id = fields.Many2one("request.medium", "Request Medium")
