+++ purchase_order_request_medium/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_request_medium_user,request_medium_user,model_request_medium,base.group_user,1,0,0,0
+access_request_medium_manager,request_medium_manager,model_request_medium,purchase.group_purchase_manager,1,1,1,1
