+++ purchase_supplier_info/__init__.py
@@ -1,0 +1,1 @@
+from . import models
