+++ purchase_supplier_info/__manifest__.py
@@ -1,0 +1,15 @@
+# Copyright 2017 Quartile Limited
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{
+    "name": "Purchase Supplier Info",
+    "version": "11.0.1.0.0",
+    "author": "Quartile Limited",
+    "website": "https://www.quartile.co",
+    "category": "Purchase",
+    "license": "AGPL",
+    "depends": ["purchase_ext_sst"],
+    "data": [
+        "views/purchase_order_views.xml",
+    ],
+    "installable": True,
+}
