+++ purchase_supplier_info/models/__init__.py
@@ -1,0 +1,1 @@
+from . import purchase_order
