+++ purchase_supplier_info/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module adds the information of supplier in purchase and update supplier's information
+when update values in purchase order.
